refactor: extract servlets from transport
